requirements.txt
# Google API Client Libraries
google-api-python-client==2.108.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0

# Additional dependencies
requests==2.31.0
